feat: Add abstract class Conteudo and subclasses Mentoria and Curso
The code changes introduce an abstract class `Conteudo` and two subclasses `Mentoria` and `Curso`. The `Conteudo` class has been modified to be abstract and includes a new static final variable `XP_PADRAO`. The subclasses `Mentoria` and `Curso` extend `Conteudo` and implement the `calcularXp()` method.

The `Mentoria` class includes a new instance variable `data` and overrides the `calcularXp()` method to return `XP_PADRAO + 20d`. It also includes a `toString()` method to display the title, description, and date of the mentoring session.

The `Curso` class includes a new instance variable `cargaHoraria` and overrides the `calcularXp()` method to return `XP_PADRAO * cargaHoraria`. It also includes a `toString()` method to display the title, description, and duration of the course.